product commit

+++ Techshop_Assignment/src/dao/ProductInterface.java
@@ -8,7 +8,7 @@ public interface ProductInterface {
     void insertProduct(Products product) throws Exception ;
     void getProductDetails(int productID) throws ClassNotFoundException;
     List<Products> getProductList();
-    void updateProductInfo(int productId, int newPrice,String newDescription) throws InvalidDataException, ClassNotFoundException;
+    void updateProductInfo(int productId, Double newPrice,String newDescription) throws InvalidDataException, ClassNotFoundException;
     void isProductInStock(int productId) throws InvalidDataException, ClassNotFoundException ;
     List<Products> searchProducts(String keyword) throws ClassNotFoundException;
     void closeResources();
